+++ concordium-node/src/lib.rs
@@ -47,3 +47,6 @@ pub mod consensus_ffi;
 pub mod test_utils;
 
 pub mod flatbuffers_shim;
+
+#[cfg(target_os = "macos")]
+mod macos_log;